refactor(channels): drop ingress exports orphaned by the whatsapp split
